Move pylxd.image.LXDImage to pylxd.deprecated.image
